+++ tests/integration_tests/integration_tests.py
@@ -1,14 +1,8 @@
 import os
-import hashlib
-import subprocess
 import pathlib
 import re
-import itertools
-import tempfile
-import sys
 
 import pytest
-import gitlab
 
 import repobee_plug as plug
 
@@ -16,454 +10,45 @@ import _repobee.ext
 import _repobee.ext.gitlab
 import _repobee.cli.mainparser
 
-import gitlabmanager
+from asserts import (
+    assert_master_repos_exist,
+    assert_repos_exist,
+    assert_repos_contain,
+    assert_on_groups,
+    assert_issues_exist,
+    assert_num_issues,
+    assert_cloned_repos,
+)
+from conftest import expected_num_members_group_assertion
+from const import (
+    VOLUME_DST,
+    BASE_DOMAIN,
+    LOCAL_DOMAIN,
+    ORG_NAME,
+    MASTER_ORG_NAME,
+    ACTUAL_USER,
+    MASTER_REPO_NAMES,
+    STUDENT_TEAMS,
+    STUDENT_TEAM_NAMES,
+    STUDENT_REPO_NAMES,
+    REPOBEE_GITLAB,
+    BASE_ARGS_NO_TB,
+    BASE_ARGS,
+    STUDENTS_ARG,
+    MASTER_REPOS_ARG,
+    MASTER_ORG_ARG,
+)
+from helpers import (
+    api_instance,
+    run_in_docker_with_coverage,
+    run_in_docker,
+    update_repo,
+    hash_directory,
+)
 
 assert os.getenv(
     "REPOBEE_NO_VERIFY_SSL"
 ), "The env variable REPOBEE_NO_VERIFY_SSL must be set to 'true'"
-
-
-VOLUME_DST = "/workdir"
-COVERAGE_VOLUME_DST = "/coverage"
-DIR = pathlib.Path(__file__).resolve().parent
-TOKEN = (DIR / "token").read_text(encoding="utf-8").strip()
-
-OAUTH_USER = "oauth2"
-BASE_DOMAIN = "gitlab.integrationtest.local"
-BASE_URL = "https://" + BASE_DOMAIN
-LOCAL_DOMAIN = "localhost:50443"
-LOCAL_BASE_URL = "https://" + LOCAL_DOMAIN
-ORG_NAME = "dd1337-fall2020"
-MASTER_ORG_NAME = "dd1337-master"
-ACTUAL_USER = "ric"
-
-MASTER_REPO_NAMES = [
-    p.name for p in (DIR / "dd1337-master-repos").iterdir() if p.is_dir()
-]
-STUDENT_TEAMS = [
-    plug.Team(members=[s.strip()])
-    for s in pathlib.Path("students.txt").read_text().strip().split("\n")
-]
-STUDENT_TEAM_NAMES = [str(t) for t in STUDENT_TEAMS]
-STUDENT_REPO_NAMES = plug.generate_repo_names(STUDENT_TEAMS, MASTER_REPO_NAMES)
-
-REPOBEE_GITLAB = "repobee -p gitlab"
-BASE_ARGS_NO_TB = ["--bu", BASE_URL, "-o", ORG_NAME, "-t", TOKEN]
-BASE_ARGS = [*BASE_ARGS_NO_TB, "--tb"]
-STUDENTS_ARG = ["-s", " ".join(STUDENT_TEAM_NAMES)]
-MASTER_REPOS_ARG = ["--mn", " ".join(MASTER_REPO_NAMES)]
-MASTER_ORG_ARG = ["--mo", MASTER_ORG_NAME]
-
-TASK_CONTENTS_SHAS = {
-    "task-1": b"\xb0\xb0,t\xd1\xe9a bu\xdfX\xcf,\x98\xd2\x04\x1a\xe8\x88",
-    "task-2": b"\x1d\xdc\xa6A\xd7\xec\xdc\xc6FSN\x01\xdf|\x95`U\xb5\xdc\x9d",
-    "task-3": b"Q\xd1x\x13r\x02\xd9\x98\xa2\xb2\xd9\xe3\xa9J^\xa2/X\xbe\x1b",
-}
-
-
-def api_instance(org_name=ORG_NAME):
-    """Return a valid instance of the GitLabAPI class."""
-    return _repobee.ext.gitlab.GitLabAPI(LOCAL_BASE_URL, TOKEN, org_name)
-
-
-def gitlab_and_groups():
-    """Return a valid gitlab instance, along with the master group and the
-    target group.
-    """
-    gl = gitlab.Gitlab(LOCAL_BASE_URL, private_token=TOKEN, ssl_verify=False)
-    master_group = gl.groups.list(search=MASTER_ORG_NAME)[0]
-    target_group = gl.groups.list(search=ORG_NAME)[0]
-    return gl, master_group, target_group
-
-
-def assert_master_repos_exist(master_repo_names, org_name):
-    """Assert that the master repos are in the specified group."""
-    gl, *_ = gitlab_and_groups()
-    group = gl.groups.list(search=org_name)[0]
-    actual_repo_names = [g.name for g in group.projects.list(all=True)]
-    assert sorted(actual_repo_names) == sorted(master_repo_names)
-
-
-def assert_repos_exist(student_teams, master_repo_names, org_name=ORG_NAME):
-    """Assert that the associated student repos exist."""
-    repo_names = plug.generate_repo_names(student_teams, master_repo_names)
-    gl = gitlab.Gitlab(LOCAL_BASE_URL, private_token=TOKEN, ssl_verify=False)
-    target_group = gl.groups.list(search=ORG_NAME)[0]
-    student_groups = gl.groups.list(id=target_group.id)
-
-    projects = [p for g in student_groups for p in g.projects.list(all=True)]
-    project_names = [p.name for p in projects]
-
-    assert set(project_names) == set(repo_names)
-
-
-def assert_repos_contain(
-    student_teams, master_repo_names, filename, text, org=ORG_NAME
-):
-    """Assert that each of the student repos contain the given file."""
-    repo_names = plug.generate_repo_names(student_teams, master_repo_names)
-    gl = gitlab.Gitlab(LOCAL_BASE_URL, private_token=TOKEN, ssl_verify=False)
-    target_group = gl.groups.list(search=ORG_NAME)[0]
-    student_groups = gl.groups.list(id=target_group.id)
-
-    projects = [
-        gl.projects.get(p.id)
-        for g in student_groups
-        for p in g.projects.list(all=True)
-        if p.name in repo_names
-    ]
-    assert len(projects) == len(repo_names)
-    for project in projects:
-        assert (
-            project.files.get(filename, "master").decode().decode("utf8")
-            == text
-        )
-
-
-def assert_on_groups(
-    student_teams,
-    org_name=ORG_NAME,
-    single_group_assertion=None,
-    all_groups_assertion=None,
-):
-    """Assert that the expected student groups exist and that the expected
-    members are in them.
-
-    single_group_assertion(expected, actual) should be a function that takes
-    the ``expected`` Team, and the ``actual`` gitlab Group. If provided, the
-    custom assertion is used INSTEAD of the default single group assertion.
-
-    all_groups_assertion(expected, actual) should be a function that takes the
-    ``expected`` teams and asserts them against all ``actual`` groups. If
-    provided, this is used INSTEAD of the default all-groups assertion.
-    """
-    gl = gitlab.Gitlab(LOCAL_BASE_URL, private_token=TOKEN, ssl_verify=False)
-    target_group = gl.groups.list(search=ORG_NAME)[0]
-    sorted_teams = sorted(list(student_teams), key=lambda t: t.name)
-    team_names = set(t.name for t in sorted_teams)
-
-    sorted_groups = sorted(
-        [
-            g
-            for g in gl.groups.list(id=target_group.id)
-            if g.name in team_names
-        ],
-        key=lambda g: g.name,
-    )
-
-    if all_groups_assertion:
-        all_groups_assertion(sorted_teams, sorted_groups)
-    else:
-        assert set(g.name for g in sorted_groups) == set(
-            str(team) for team in sorted_teams
-        )
-    for group, team in zip(sorted_groups, sorted_teams):
-        # the user who owns the OAUTH token is always listed as a member
-        # of groups he/she creates
-        if single_group_assertion is None:
-            expected_members = sorted(team.members + [ACTUAL_USER])
-            actual_members = sorted(m.username for m in group.members.list(all=True))
-            assert group.name == team.name
-            assert actual_members == expected_members
-        else:
-            single_group_assertion(expected=team, actual=group)
-
-
-def _assert_on_projects(student_teams, master_repo_names, assertion):
-    """Execute the specified assertion operation on a project. Assertion should
-    be a callable taking precisely on project as an argument.
-    """
-    gl = gitlab.Gitlab(LOCAL_BASE_URL, private_token=TOKEN, ssl_verify=False)
-    repo_names = plug.generate_repo_names(student_teams, master_repo_names)
-    target_group = gl.groups.list(search=ORG_NAME)[0]
-    student_groups = gl.groups.list(id=target_group.id)
-    projects = [
-        gl.projects.get(p.id)
-        for g in student_groups
-        for p in g.projects.list(all=True)
-        if p.name in repo_names
-    ]
-
-    for proj in projects:
-        assertion(proj)
-
-
-def assert_issues_exist(
-    student_teams,
-    master_repo_names,
-    expected_issue,
-    expected_state="opened",
-    expected_num_asignees=0,
-):
-    """Assert that the expected issue has been opened in each of the student
-    repos.
-    """
-
-    def assertion(project):
-        issues = project.issues.list(all=True)
-        for actual_issue in issues:
-            if actual_issue.title == expected_issue.title:
-                assert actual_issue.state == expected_state
-                assert actual_issue.description == expected_issue.body
-                assert len(actual_issue.assignees) == expected_num_asignees
-                assert ACTUAL_USER not in [
-                    asignee["username"] for asignee in actual_issue.assignees
-                ]
-
-                return
-        assert False, "no issue matching the specified title"
-
-    _assert_on_projects(student_teams, master_repo_names, assertion)
-
-
-def assert_num_issues(student_teams, master_repo_names, num_issues):
-    """Assert that there are precisely num_issues issues in each student
-    repo.
-    """
-
-    def assertion(project):
-        assert len(project.issues.list(all=True)) == num_issues
-
-    _assert_on_projects(student_teams, master_repo_names, assertion)
-
-
-@pytest.fixture(autouse=True, scope="session")
-def setup_gitlab_instance():
-    """Perform first-time setup of the GitLab instance."""
-    gitlabmanager.setup()
-
-
-@pytest.fixture(autouse=True, scope="session")
-def teardown_gitlab_instance():
-    """Teardown the GitLab instance after all tests have finished."""
-    yield
-    gitlabmanager.teardown()
-
-
-@pytest.fixture(autouse=True)
-def restore():
-    """Run the script that restores the GitLab instance to its initial
-    state.
-    """
-    gitlabmanager.restore()
-
-
-@pytest.fixture
-def tmpdir_volume_arg(tmpdir):
-    """Create a temporary directory and return an argument string that
-    will mount a docker volume to it.
-    """
-    yield "-v {}:{}".format(str(tmpdir), VOLUME_DST)
-
-
-@pytest.fixture(scope="module", autouse=True)
-def coverage_volume():
-    covdir = pathlib.Path(".").resolve() / ".coverage_files"
-    yield "-v {}:{}".format(str(covdir), COVERAGE_VOLUME_DST)
-    covfile = covdir / ".coverage"
-    assert covfile.is_file()
-
-
-@pytest.fixture(scope="module", autouse=True)
-def generate_coverage_report(coverage_volume):
-    """Generate a coverage report after all tests have run."""
-    yield
-    # xml report for Codecov
-    run_in_docker(
-        "cd {} && coverage xml".format(COVERAGE_VOLUME_DST),
-        extra_args=[coverage_volume],
-    )
-    # txt report for manual inspection
-    run_in_docker(
-        "cd {} && coverage report > report.txt".format(COVERAGE_VOLUME_DST),
-        extra_args=[coverage_volume],
-    )
-
-
-@pytest.fixture(autouse=True)
-def handle_coverage_file(extra_args):
-    """Copy the coverage file back and forth."""
-    # copy the previous .coverage file into the workdir
-    run_in_docker(
-        "cp {}/.coverage .".format(COVERAGE_VOLUME_DST), extra_args=extra_args
-    )
-    yield
-    # copy the appended .coverage file into the coverage volume
-    run_in_docker(
-        "cp .coverage {}".format(COVERAGE_VOLUME_DST), extra_args=extra_args
-    )
-
-
-@pytest.fixture
-def extra_args(tmpdir_volume_arg, coverage_volume):
-    """Extra arguments to pass to run_in_docker when executing a test."""
-    return [tmpdir_volume_arg, coverage_volume]
-
-
-def run_in_docker_with_coverage(command, extra_args=None):
-    assert extra_args, "extra volume args are required to run with coverage"
-    coverage_command = (
-        "coverage run --branch --append --source _repobee -m " + command
-    )
-    return run_in_docker(coverage_command, extra_args=extra_args)
-
-
-def run_in_docker(command, extra_args=None):
-    extra_args = " ".join(extra_args) if extra_args else ""
-    docker_command = (
-        "sudo docker run {} --net development --rm --name repobee "
-        "repobee:test /bin/sh -c '{}'"
-    ).format(extra_args, command)
-    proc = subprocess.run(
-        docker_command,
-        shell=True,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,
-    )
-    print(proc.stdout.decode(sys.getdefaultencoding())) # for test output on failure
-    return proc
-
-
-def update_repo(repo_name, filename, text):
-    """Add a file with the given filename and text to the repo."""
-    gl = gitlab.Gitlab(LOCAL_BASE_URL, private_token=TOKEN, ssl_verify=False)
-    proj, *_ = [
-        p for p in gl.projects.list(search=repo_name) if p.name == repo_name
-    ]
-    env = os.environ.copy()
-    env["GIT_SSL_NO_VERIFY"] = "true"
-    env["GIT_AUTHOR_EMAIL"] = "slarse@kth.se"
-    env["GIT_AUTHOR_NAME"] = "Simon"
-    env["GIT_COMMITTER_EMAIL"] = env["GIT_AUTHOR_EMAIL"]
-    env["GIT_COMMITTER_NAME"] = env["GIT_AUTHOR_NAME"]
-    with tempfile.TemporaryDirectory() as tmpdir:
-        url_with_token = (
-            proj.web_url.replace(
-                "https://", "https://oauth2:{}@".format(TOKEN)
-            ).replace(BASE_DOMAIN, LOCAL_DOMAIN)
-            + ".git"
-        )
-        clone_proc = subprocess.run(
-            "git clone {}".format(url_with_token).split(), cwd=tmpdir, env=env
-        )
-        assert clone_proc.returncode == 0
-
-        repo_dir = pathlib.Path(tmpdir) / proj.name
-        new_file = repo_dir / filename
-        new_file.touch()
-        new_file.write_text(text)
-
-        add_proc = subprocess.run(
-            "git add {}".format(filename).split(), cwd=str(repo_dir), env=env
-        )
-        assert add_proc.returncode == 0
-
-        commit_proc = subprocess.run(
-            "git commit -am newfile".split(), cwd=str(repo_dir), env=env
-        )
-        assert commit_proc.returncode == 0
-
-        push_proc = subprocess.run(
-            "git push".split(), cwd=str(repo_dir), env=env
-        )
-        assert push_proc.returncode == 0
-
-    assert proj.files.get(filename, "master").decode().decode("utf8") == text
-
-
-@pytest.fixture
-def with_student_repos(restore):
-    """Set up student repos before starting tests.
-
-    Note that explicitly including restore here is necessary to ensure that
-    it runs before this fixture.
-    """
-    command = " ".join(
-        [
-            REPOBEE_GITLAB,
-            _repobee.cli.mainparser.SETUP_PARSER,
-            *BASE_ARGS,
-            *MASTER_ORG_ARG,
-            *MASTER_REPOS_ARG,
-            *STUDENTS_ARG,
-        ]
-    )
-
-    result = run_in_docker(command)
-
-    # pre-test asserts
-    assert result.returncode == 0
-    assert_repos_exist(STUDENT_TEAMS, MASTER_REPO_NAMES)
-    assert_on_groups(STUDENT_TEAMS)
-
-
-def assert_cloned_repos(repo_names, tmpdir):
-    """Check that the cloned repos have the expected contents.
-
-    NOTE: Only checks the contents of the root of the project.
-    """
-    # group by master repo name, all of which have the same length
-    grouped_repo_names = itertools.groupby(
-        sorted(repo_names),
-        key=lambda name: name[len(name) - len(MASTER_REPO_NAMES[0]) :],
-    )
-
-    root = pathlib.Path(tmpdir).resolve()
-
-    for master_repo_name, student_repo_names in grouped_repo_names:
-
-        expected_sha = TASK_CONTENTS_SHAS[master_repo_name]
-        for repo_name in student_repo_names:
-            sha = hash_directory(root / repo_name)
-            assert sha == expected_sha
-            assert (root / repo_name / ".git").is_dir()
-
-
-def hash_directory(path):
-    shas = []
-    for dirpath, _, filenames in os.walk(str(path)):
-        if ".git" in dirpath:
-            continue
-        files = list(
-            pathlib.Path(dirpath) / filename for filename in filenames
-        )
-        shas += (hashlib.sha1(file.read_bytes()).digest() for file in files)
-    return hashlib.sha1(b"".join(sorted(shas))).digest()
-
-
-@pytest.fixture
-def open_issues(with_student_repos):
-    """Open two issues in each student repo."""
-    task_issue = plug.Issue(
-        title="Task", body="The task is to do this, this and that"
-    )
-    correction_issue = plug.Issue(
-        title="Correction required", body="You need to fix this, this and that"
-    )
-    issues = [task_issue, correction_issue]
-    gl = gitlab.Gitlab(LOCAL_BASE_URL, private_token=TOKEN, ssl_verify=False)
-    target_group = gl.groups.list(search=ORG_NAME)[0]
-    projects = (
-        gl.projects.get(p.id)
-        for p in target_group.projects.list(include_subgroups=True, all=True)
-    )
-    for project in projects:
-        project.issues.create(
-            dict(title=task_issue.title, description=task_issue.body)
-        )
-        project.issues.create(
-            dict(
-                title=correction_issue.title, description=correction_issue.body
-            )
-        )
-
-    assert_num_issues(STUDENT_TEAM_NAMES, MASTER_REPO_NAMES, len(issues))
-    assert_issues_exist(STUDENT_TEAM_NAMES, MASTER_REPO_NAMES, task_issue)
-    assert_issues_exist(
-        STUDENT_TEAM_NAMES, MASTER_REPO_NAMES, correction_issue
-    )
-
-    return issues
 
 
 @pytest.mark.filterwarnings("ignore:.*Unverified HTTPS request.*")
@@ -868,23 +453,6 @@ class TestListIssues:
             assert not re.search(unexpected_pattern, output, search_flags)
 
 
-def expected_num_members_group_assertion(expected_num_members):
-    def group_assertion(expected, actual):
-        assert expected.name == actual.name
-        # +1 member for the group owner
-        assert len(actual.members.list(all=True)) == expected_num_members + 1
-        assert len(actual.projects.list(all=True)) == 1
-        project_name = actual.projects.list(all=True)[0].name
-        assert actual.name.startswith(project_name)
-        for member in actual.members.list(all=True):
-            if member.username == ACTUAL_USER:
-                continue
-            assert member.username not in project_name
-            assert member.access_level == gitlab.REPORTER_ACCESS
-
-    return group_assertion
-
-
 @pytest.mark.filterwarnings("ignore:.*Unverified HTTPS request.*")
 class TestAssignReviews:
     """Tests for the assign-reviews command."""
@@ -965,41 +533,6 @@ class TestAssignReviews:
         )
         assert result.returncode == 1
         assert_num_issues(STUDENT_TEAMS, [master_repo_name], 0)
-
-
-@pytest.fixture
-def with_reviews(with_student_repos):
-    master_repo_name = MASTER_REPO_NAMES[1]
-    expected_review_teams = [
-        plug.Team(
-            members=[],
-            name=plug.generate_review_team_name(
-                student_team_name, master_repo_name
-            ),
-        )
-        for student_team_name in STUDENT_TEAM_NAMES
-    ]
-    command = " ".join(
-        [
-            REPOBEE_GITLAB,
-            _repobee.cli.mainparser.ASSIGN_REVIEWS_PARSER,
-            *BASE_ARGS,
-            "--mn",
-            master_repo_name,
-            *STUDENTS_ARG,
-            "-n",
-            "1",
-        ]
-    )
-
-    result = run_in_docker(command)
-
-    assert result.returncode == 0
-    assert_on_groups(
-        expected_review_teams,
-        single_group_assertion=expected_num_members_group_assertion(1),
-    )
-    return (master_repo_name, expected_review_teams)
 
 
 @pytest.mark.filterwarnings("ignore:.*Unverified HTTPS request.*")
